Added unit tests to cover edge cases
